picker improvements
